Don't guess `ClientProvider` from feature flags in iroh-relay, deduplicate `NoCertProvider`
