+++ iroh-relay/src/tls.rs
@@ -141,7 +141,7 @@ impl CaRootsConfig {
             }
             #[cfg(any(test, feature = "test-utils"))]
             Mode::InsecureSkipVerify => {
-                Arc::new(self::no_cert_verifier::NoCertVerifier { crypto_provider })
+                Arc::new(no_cert_verifier::NoCertVerifier { crypto_provider })
             }
         })
     }
@@ -173,6 +173,25 @@ pub fn default_provider() -> Arc<CryptoProvider> {
 #[cfg(all(feature = "tls-aws-lc-rs", not(feature = "tls-ring")))]
 pub fn default_provider() -> Arc<CryptoProvider> {
     Arc::new(rustls::crypto::aws_lc_rs::default_provider())
+}
+
+#[cfg(all(any(test, feature = "test-utils"), with_crypto_provider))]
+/// Creates a client config that trusts any servers without verifying their TLS certificate.
+///
+/// Should be used for testing local relay setups only.
+pub fn make_dangerous_client_config() -> rustls::ClientConfig {
+    tracing::warn!(
+        "Insecure config: SSL certificates from relay servers will be trusted without verification"
+    );
+    let crypto_provider = crate::tls::default_provider();
+    rustls::client::ClientConfig::builder_with_provider(crypto_provider.clone())
+        .with_protocol_versions(&[&rustls::version::TLS13])
+        .expect("protocols supported by ring")
+        .dangerous()
+        .with_custom_certificate_verifier(Arc::new(no_cert_verifier::NoCertVerifier {
+            crypto_provider,
+        }))
+        .with_no_client_auth()
 }
 
 #[cfg(any(test, feature = "test-utils"))]
